Fix compiler warning.
git-svn-id: http://ext3grep.googlecode.com/svn/trunk@17 bd4756ed-6648-0410-a231-31fef783fe29

+++ src/locate.cc
@@ -170,7 +170,7 @@ std::string parent_directory(int blocknr, std::set<std::string> const& filenames
 #endif
     };
     regex_t preg;
-    for (unsigned int re = 0; re < sizeof(table) / sizeof(table[0]); ++re)
+    for (size_t re = 0; re < sizeof(table) / sizeof(table[0]); ++re)
     {
       int hits = 0;
       regcomp(&preg, table[re].regexp, REG_EXTENDED|REG_NOSUB);
